feat: rename MathingGame to MatchingGame
