+++ MatchingGame.Android/MathingGame.Android.csproj
@@ -7,8 +7,8 @@
     <ProjectTypeGuids>{EFBA0AD7-5A72-4C68-AF49-83D382785DCF};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <TemplateGuid>{c9e5eea5-ca05-42a1-839b-61506e0a37df}</TemplateGuid>
     <OutputType>Library</OutputType>
-    <RootNamespace>MathingGame.Droid</RootNamespace>
-    <AssemblyName>MathingGame.Android</AssemblyName>
+    <RootNamespace>MatchingGame.Droid</RootNamespace>
+    <AssemblyName>MatchingGame.Android</AssemblyName>
     <Deterministic>True</Deterministic>
     <AndroidApplication>True</AndroidApplication>
     <AndroidResgenFile>Resources\Resource.designer.cs</AndroidResgenFile>
@@ -87,9 +87,9 @@
     <Folder Include="Resources\drawable\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MathingGame\MathingGame.csproj">
+    <ProjectReference Include="..\MatchingGame\MatchingGame.csproj">
       <Project>{9501C3E1-B1E2-4251-AB22-BEC7287D1164}</Project>
-      <Name>MathingGame</Name>
+      <Name>MatchingGame</Name>
     </ProjectReference>
   </ItemGroup>
  <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
