+++ MatchingGame.Android/Resources/mipmap-anydpi-v26/icon.xml
