+++ MatchingGame.Android/Resources/mipmap-anydpi-v26/icon_round.xml
