+++ MatchingGame.Android/Resources/mipmap-hdpi/launcher_foreground.png
