+++ MatchingGame.Android/Resources/mipmap-mdpi/launcher_foreground.png
