+++ MatchingGame.Android/Resources/mipmap-xhdpi/launcher_foreground.png
