+++ MatchingGame.Android/Resources/mipmap-xxhdpi/launcher_foreground.png
