+++ MatchingGame.Android/Resources/mipmap-xxxhdpi/launcher_foreground.png
