+++ MatchingGame.Android/Resources/values/colors.xml
