+++ MatchingGame.Android/Resources/values/styles.xml
