+++ MatchingGame.iOS/Assets.xcassets/AppIcon.appiconset/Contents.json
