+++ MatchingGame.iOS/Info.plist
@@ -23,15 +23,15 @@
     <key>MinimumOSVersion</key>
     <string>8.0</string>
     <key>CFBundleDisplayName</key>
-    <string>MathingGame</string>
+    <string>MatchingGame</string>
     <key>CFBundleIdentifier</key>
-    <string>com.companyname.MathingGame</string>
+    <string>com.companyname.MatchingGame</string>
     <key>CFBundleVersion</key>
     <string>1.0</string>
     <key>UILaunchStoryboardName</key>
     <string>LaunchScreen</string>
     <key>CFBundleName</key>
-    <string>MathingGame</string>
+    <string>MatchingGame</string>
     <key>XSAppIconAssets</key>
     <string>Assets.xcassets/AppIcon.appiconset</string>
 </dict>
