+++ MatchingGame.iOS/MathingGame.iOS.csproj
@@ -9,9 +9,9 @@
     <ProjectTypeGuids>{FEACFBD2-3405-455C-9665-78FE426C6842};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <TemplateGuid>{6143fdea-f3c2-4a09-aafa-6e230626515e}</TemplateGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>MathingGame.iOS</RootNamespace>
+    <RootNamespace>MatchingGame.iOS</RootNamespace>
     <IPhoneResourcePrefix>Resources</IPhoneResourcePrefix>
-    <AssemblyName>MathingGame.iOS</AssemblyName>
+    <AssemblyName>MatchingGame.iOS</AssemblyName>
     <MtouchEnableSGenConc>true</MtouchEnableSGenConc>
     <MtouchHttpClientHandler>NSUrlSessionHandler</MtouchHttpClientHandler>
     <ProvisioningType>automatic</ProvisioningType>
@@ -128,9 +128,9 @@
   </ItemGroup>
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\iOS\Xamarin.iOS.CSharp.targets" />
   <ItemGroup>
-    <ProjectReference Include="..\MathingGame\MathingGame.csproj">
+    <ProjectReference Include="..\MatchingGame\MatchingGame.csproj">
       <Project>{9501C3E1-B1E2-4251-AB22-BEC7287D1164}</Project>
-      <Name>MathingGame</Name>
+      <Name>MatchingGame</Name>
     </ProjectReference>
   </ItemGroup>
 </Project>
