+++ MatchingGame/App.xaml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <Application xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MathingGame.App">
+             x:Class="MatchingGame.App">
     <Application.Resources>
 
     </Application.Resources>
